Moved collections.py -> entry_collections.py to avoid nameclash with standard library
